feat: refactor feedback components
Move feedback components to dedicated directories:
- Banner component to feedback/banner
- Cookie consent component to feedback/cookie-consent
- Updated directory exports and README.md
- Added backward compatibility exports
- Updated COMPONENT-MOVEMENT-PLAN.md
