+++ src/components/ui/feedback/banner/index.ts
@@ -1,0 +1,1 @@
+export * from "./banner";
