+++ src/components/ui/feedback/cookie-consent/index.ts
@@ -1,0 +1,1 @@
+export * from "./cookie-consent";
