+++ src/components/ui/feedback/index.ts
@@ -1,3 +1,6 @@
 export * from './toast';
 export * from './toaster';
 export * from './alert';
+export * from './banner';
+export * from './cookie-consent';
+export * from './tooltip';
